+++ scripts/__init__.py
@@ -1,3 +1,3 @@
 from .start_conky import StartConky
-from .system import SystemConkyConf
-from .info import InfoConkyConf
+from .system_conky_conf import SystemConkyConf
+from .info_conky_conf import InfoConkyConf
