+++ scripts/conky_conf_writer.py
